requirements.txt
nintendoclients==0.0.5
anyio==3.6.2